prototypes/BeyondPresence/pyproject.toml
[project]
name = "beyondpresence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "livekit-agents>=1.1.1",
    "livekit-plugins-bey>=1.1.1",
    "livekit-plugins-openai>=1.1.1",
]
